// frontend/src/api/Api.ts
export interface TariffItem {
  /** Id */
  id: number;
  /**
   * Название
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  /** Статус */
  status?: 1 | 2;
  /** Price */
  price: number;
  /**
   * RAM
   * @min -2147483648
   * @max 2147483647
   */
  ram: number;
  /**
   * CPU
   * @min -2147483648
   * @max 2147483647
   */
  cpu: number;
  /**
   * SSD
   * @min -2147483648
   * @max 2147483647
   */
  ssd: number;
  /**
   * Image
   * @format uri
   */
  image?: string;
  /** Count */
  count?: number;
}

export interface Forecast {
  /** Id */
  id: number;
  /** Owner */
  owner?: string;
  /** Moderator */
  moderator?: string;
  tariffs?: TariffItem[];
  /** Статус */
  status?: 1 | 2 | 3 | 4 | 5;
  /**
   * Дата создания
   * @format date-time
   */
  date_created?: string;
  /**
   * Дата формирования
   * @format date-time
   */
  date_formation?: string | null;
  /**
   * Дата завершения
   * @format date-time
   */
  date_complete?: string | null;
  /**
   * Days
   * @min -2147483648
   * @max 2147483647
   */
  days?: number | null;
  /**
   * Price
   * @min -2147483648
   * @max 2147483647
   */
  price?: number | null;
}